[b-7] changes after review; add RestTemplate

+++ movieland-service/src/main/java/com/art/movieland/service/CurrencyService.java
@@ -3,5 +3,5 @@ package com.art.movieland.service;
 import com.art.movieland.entity.Currency;
 
 public interface CurrencyService {
-    float getRate(Currency currency);
+    double getRate(Currency currency);
 }
